--- packages/manifest/src/functions/is-dereferenced-board.ts
@@ -1,8 +1,0 @@
-import { DereferencedBoard } from "../types/boards";
-import { isBglLike } from "./is-bgl-like";
-
-export function isDereferencedBoard(
-  resource: object
-): resource is DereferencedBoard {
-  return isBglLike(resource);
-}
