--- packages/manifest/src/functions/is-local-uri.ts
@@ -1,5 +1,0 @@
-import { isRemoteUri } from "./is-remote-uri";
-
-export function isLocalUri(uri: string): boolean {
-  return !isRemoteUri(uri);
-}
