--- packages/manifest/src/functions/is-remote-uri.ts
@@ -1,8 +1,0 @@
-export function isRemoteUri(uri: string): boolean {
-  try {
-    new URL(uri);
-    return true;
-  } catch (e) {
-    return false;
-  }
-}
