+++ packages/manifest/src/scripts/util/ascend-to-package-dir.ts
