+++ packages/manifest/src/types/guards/index.ts
@@ -1,0 +1,4 @@
+export * from "./board-resource";
+export * from "./manifest-resource";
+export * from "./resource-reference";
+
