+++ packages/manifest/src/types/index.ts
@@ -4,3 +4,9 @@
 export type AdditionalProperties = {
   [x: string | number | symbol]: unknown;
 };
+
+export * from "./boards";
+export * from "./guards";
+export * from "./manifest";
+export * from "./resource";
+
